requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
Pillow
django-filter
drf-yasg
celery[redis]
